Added password check before deleting an item.

+++ controllers/deleteController.js
@@ -1,21 +1,75 @@
+require('dotenv').config();
+const db = require('../db/getQueries');
 const dbDelete = require('../db/deleteQueries');
+const links = require('../links');
+const { body, validationResult } = require('express-validator');
 
-async function deleteGame(req, res) {
+async function confirmDeleteGame(req, res) {
+  const title = 'game';
+  const [ selectedGame ] = await db.getSingleGameFromDB(req.params.id);
+
+  res.render('deleteConfirmation', { links, title, selectedGame });
+}
+
+async function confirmDeleteGenre(req, res) {
+  const title = 'genre';
+  const [ selectedGenre ] = await db.getSingleGenreFromDB(req.params.genre_id);
+
+  res.render('deleteConfirmation', { links, title, selectedGenre });
+}
+
+async function confirmDeleteDeveloper(req, res) {
+  const title = 'developer';
+  const [ selectedDeveloper ] = await db.getSingleDeveloperFromDB(req.params.developer_id);
+
+  res.render('deleteConfirmation', { links, title, selectedDeveloper });
+}
+
+const validatePassword = [
+  body('password').trim()
+    .equals(process.env.DELETE_PASSWORD).withMessage(`Incorrect password`)
+];
+
+const deleteGame = [ validatePassword, async (req, res) => {
+  const title = 'game';
+  const [ selectedGame ] = await db.getSingleGameFromDB(req.params.id);
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).render('deleteConfirmation', { links, title, selectedGame, errors: errors.array() });
+  }
+
   await dbDelete.deleteGameFromDB(req.params.id);
 
   res.redirect('/');
-}
+}];
 
-async function deleteGenre(req, res) {
+const deleteGenre = [ validatePassword, async (req, res) => {
+  const title = 'genre';
+  const [ selectedGenre ] = await db.getSingleGenreFromDB(req.params.genre_id);
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).render('deleteConfirmation', { links, title, selectedGenre, errors: errors.array() });
+  }
+
   await dbDelete.deleteGenreFromDB(req.params.genre_id);
 
   res.redirect('/genres');
-}
+}];
 
-async function deleteDeveloper(req, res) {
+const deleteDeveloper = [ validatePassword, async (req, res) => {
+  const title = 'developer';
+  const [ selectedDeveloper ] = await db.getSingleDeveloperFromDB(req.params.developer_id);
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).render('deleteConfirmation', { links, title, selectedDeveloper, errors: errors.array() });
+  }
+
   await dbDelete.deleteDeveloperFromDB(req.params.developer_id);
 
   res.redirect('/developers');
-}
+}];
 
-module.exports = { deleteGame, deleteGenre, deleteDeveloper };
+module.exports = { confirmDeleteGame, confirmDeleteGenre, confirmDeleteDeveloper, deleteGame, deleteGenre, deleteDeveloper };
